+++ src/bias/MovingRestraint.cpp
@@ -100,7 +100,7 @@ the total bias deposited
 class MovingRestraint : public Bias {
   std::vector<std::vector<double> > at;
   std::vector<std::vector<double> > kappa;
-  std::vector<long int> step;
+  std::vector<long long int> step;
   std::vector<double> oldaa;
   std::vector<double> oldk;
   std::vector<double> olddpotdk;
@@ -152,7 +152,7 @@ MovingRestraint::MovingRestraint(const ActionOptions&ao):
   verse(getNumberOfArguments())
 {
   parseVector("VERSE",verse);
-  std::vector<long int> ss(1); ss[0]=-1;
+  std::vector<long long int> ss(1); ss[0]=-1;
   std::vector<double> kk( getNumberOfArguments() ), aa( getNumberOfArguments() );
   for(int i=0;; i++) {
     // Read in step
@@ -173,7 +173,7 @@ MovingRestraint::MovingRestraint(const ActionOptions&ao):
   checkRead();
 
   for(unsigned i=0; i<step.size(); i++) {
-    log.printf("  step%u %ld\n",i,step[i]);
+    log.printf("  step%u %lld\n",i,step[i]);
     log.printf("  at");
     for(unsigned j=0; j<at[i].size(); j++) log.printf(" %f",at[i][j]);
     log.printf("\n");
@@ -209,7 +209,7 @@ void MovingRestraint::calculate() {
   double ene=0.0;
   double totf2=0.0;
   unsigned narg=getNumberOfArguments();
-  long int now=getStep();
+  long long int now=getStep();
   std::vector<double> kk(narg),aa(narg),f(narg),dpotdk(narg);
   if(now<=step[0]) {
     kk=kappa[0];
